+++ metadata.rb
@@ -8,8 +8,8 @@ supports 'ubuntu', ">= 14.04"
 supports 'mac_os_x'
 supports 'rhel'
 
-version          '0.1.3'
+version          '0.1.4'
 
 depends 'apt', '>= 2.4'
-depends 'ark', '~> 0.9.0'
+depends 'ark', '>= 0.9.0'